requirements.txt
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
praw>=7.7.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
alpaca-py>=0.43.0
